+++ cypress/support/gutenberg/commands/attributes.js
@@ -1,7 +1,7 @@
 /**
- * Internal dependencies
+ * External dependencies
  */
-import { dispatchResolver } from '../util'
+import { dispatchResolver } from '~common/util'
 
 /**
  * Register functions to Cypress Commands.
@@ -12,11 +12,14 @@ Cypress.Commands.add( 'setBlockAttribute', setBlockAttribute )
  * Command for setting attributes.
  *
  * @param {Object} attributes
+ * @param {string} clientId
  */
-export function setBlockAttribute( attributes = {} ) {
+export function setBlockAttribute( attributes = {}, clientId = '' ) {
 	cy.wp().then( wp => {
 		return new Cypress.Promise( ( resolve, reject ) => {
-			const { clientId = '' } = wp.data.select( 'core/block-editor' ).getSelectedBlock() || {}
+			if ( ! clientId ) {
+				clientId = ( wp.data.select( 'core/block-editor' ).getSelectedBlock() || {} ).clientId
+			}
 			if ( clientId ) {
 				wp.data.dispatch( 'core/block-editor' ).updateBlockAttributes( clientId, attributes )
 					.then( dispatchResolver( resolve ) )
