+++ cypress/support/gutenberg/commands/index.js
@@ -1,0 +1,9 @@
+/**
+ * Custom commands for Gutenberg.
+ */
+import './editor'
+import './blocks'
+import './inspector'
+import './attributes'
+import './controls'
+import './assertions'
