+++ cypress/support/gutenberg/index.js
@@ -1,0 +1,1 @@
+import './commands'
