+++ cypress/support/gutenberg/plugins/index.js
@@ -1,0 +1,4 @@
+export {
+	default as BlockSnapshots,
+	registerBlockSnapshots,
+} from './blockSnapshots'
