+++ cypress/support/index.js
@@ -1,9 +1,6 @@
-Cypress.on( 'window:before:load', win => {
-	// Allow the use of cypress experimental `fetch` polyfills.
-	delete win.fetch
-} )
-
-import './commands/index'
+import './wordpress'
+import './gutenberg'
+import './stackable'
 
 // Use Jest assertions
 import 'cypress-jest-adapter'
