+++ cypress/support/wordpress/commands/index.js
@@ -1,0 +1,6 @@
+/**
+ * Custom commands for WordPress.
+ */
+
+import './setup'
+import './plugins'
